Fix parsing of FLAC files with ID3 header
This improves upon the change in b6f403d. The change mistakenly seek'd
over the FLAC header once too many in case of an ID3 tag being present.
The provided test case didn't detect this because the FLAC header in
there was empty. The test file was updated.
Additionally, it also failed to seek the file to the end of the ID3
header properly in case of padding. Two test files for this case have
been added.